% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{download_all_daily_dumps}
\alias{download_all_daily_dumps}
\title{Download all available data}
\usage{
download_all_daily_dumps(base.data.directory = system.file("extdata",
  "market-data", mustWork = TRUE, package = "bitcoinchartsr"),
  overwrite = TRUE, debug = TRUE)
}
\arguments{
\item{base.data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}

\item{debug}{logical. Debugging flag.}
}
\description{
Download all available data for all markets in one fell swoop. Caution! Some exchange data files are > 500 MB!!!
}
\examples{
\dontrun{
# Download all available market data in one fell swoop:
download_all_daily_dumps()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://api.bitcoincharts.com/v1/csv/}}
}

