% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{download_all_data_files}
\alias{download_all_data_files}
\title{Download all available archived trade data}
\usage{
download_all_data_files(data.directory = system.file("extdata", mustWork =
  TRUE, package = "bitcoinchartsr"), overwrite = TRUE)
}
\arguments{
\item{data.directory}{character. Destination directory for downloaded
data files. Defaults to package install extdata/marketdata directory.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}
}
\description{
Download all available data for all markets in one fell swoop.
Caution! Some exchange data files are > 500 MB!!!
}
\examples{
\dontrun{
# Download all available market data in one fell swoop:
downdownload_all_data_files()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://api.bitcoincharts.com/v1/csv/}}
}

