% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{download_data_file}
\alias{download_data_file}
\title{Download single exchange data file from bitcoincharts.com API}
\usage{
download_data_file(symbol, data.directory = file.path(system.file("extdata",
  mustWork = TRUE, package = "bitcoinchartsr")), overwrite = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the
\code{'get_symbol_listing()'} method.}

\item{data.directory}{character. Destination directory for downloaded data files.
Defaults to package install extdata for the symbol specified directory.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}

\item{debug}{logical. Debugging flag.}
}
\description{
Download single exchange data file from bitcoincharts.com API
}
\examples{
\dontrun{
# Get the full dump of all MTGOX USD transactions
download_data_file(symbol='mtgoxUSD', data.directory='/home/user/Desktop')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://api.bitcoincharts.com/v1/csv/}}
}

