% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_all_time_trade_count}
\alias{get_all_time_trade_count}
\title{Get total number of trades ever executed on a given exchange}
\usage{
get_all_time_trade_count(symbol, base.data.directory = system.file("extdata",
  "market-data", mustWork = TRUE, package = "bitcoinchartsr"), debug = FALSE)
}
\arguments{
\item{symbol}{character. Symbol to get total trade count for}

\item{base.data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}
}
\description{
Get total number of trades ever executed on a given exchange by counting lines in dump file
}
\examples{
\dontrun{
# Download all available market data in one fell swoop:
download_all_daily_dumps()
}
}

