% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_bitcoincharts_data}
\alias{get_bitcoincharts_data}
\title{Query bitcoincharts.com API}
\usage{
get_bitcoincharts_data(symbol, start.date = as.character(as.Date(now() -
  days(30))), end.date = as.character(as.Date(now() + days(1))),
  ohlc.frequency = "hours", align = TRUE, fill = FALSE,
  data.directory = file.path(system.file("extdata", "market-data", mustWork =
  TRUE, package = "bitcoinchartsr"), symbol), download.data = FALSE,
  overwrite = FALSE, auto.assign = FALSE, env = .GlobalEnv,
  debug = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the \code{'get_symbol_listing()'} method.}

\item{start.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format representing the start of the requested data series. Defaults to 30 days prior to current date.}

\item{end.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format representing the start of the requested data series. Defaults to current date + 1.}

\item{ohlc.frequency}{character. Supported values are \code{seconds}, \code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{align}{logical. Align time series index.}

\item{fill}{logical. Fill missing values.}

\item{data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}

\item{download.data}{logical. Whether to download a fresh copy of the data file.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}

\item{auto.assign}{logical. Whether or not to auto-assign the variable to the environment specified in the \code{env} param}

\item{env}{character. Environment to auto.assign the return value to. Defaults to .GlobalEnv}

\item{debug}{logical. Debugging flag.}
}
\description{
Query bitcoincharts.com API to obtain market data for bitcoin exchanges
}
\examples{
\dontrun{
# Get one month of hourly market data for virtexCAD:
get_bitcoincharts_data('virtexCAD')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}

