% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_current_difficulty}
\alias{get_current_difficulty}
\title{Get current network difficulty from site header}
\usage{
get_current_difficulty()
}
\description{
Get current network difficulty from site header
}
\examples{
\dontrun{
# Get all market symbols:
get_current_difficulty()
}
}
\references{
\url{http://bitcoincharts.com}
}

