% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_exchange_info}
\alias{get_exchange_info}
\title{Get detailed information for a specified exchange}
\usage{
get_exchange_info(symbol, debug = FALSE)
}
\arguments{
\item{symbol}{character. Exchange you wish to obtain info for}

\item{debug}{logical. Debugging flag.}
}
\description{
Get detailed listing of all currently available exchanges from http://bitcoincharts.com/markets
}
\examples{
\dontrun{
# Get all market symbols:
get_exchange_info()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://bitcoincharts.com/markets}}
}

