% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_markets_snapshot}
\alias{get_markets_snapshot}
\title{Get snapshots of all markets listed on the site}
\usage{
get_markets_snapshot()
}
\description{
Get snapshots of all markets listed on the site
}
\examples{
\dontrun{
# Get all market symbols:
get_markets_snapshot()
}
}
\references{
\url{http://bitcoincharts.com}
}

