% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_most_recent_ohlc}
\alias{get_most_recent_ohlc}
\title{Get OHLC for last two days of trades via API}
\usage{
get_most_recent_ohlc(symbol, start.date = as.character(as.Date(now() -
  days(2))), data.directory = system.file("extdata", "market-data", mustWork =
  TRUE, package = "bitcoinchartsr"), ohlc.frequency = "hours", align = TRUE,
  fill = FALSE, debug = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the \code{'get_symbol_listing()'} method.}

\item{data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}

\item{ohlc.frequency}{character. Supported values are \code{seconds}, \code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{align}{logical. Align time series index.}

\item{fill}{logical. Fill missing values.}

\item{debug}{logical. Debugging flag.}
}
\description{
Utility function which returns the most recent OHLC candle from the bitcoincharts.com API
}
\examples{
\dontrun{
# Get most recent 1-day candle from cavirtex.com:
get_most_recent_trade('virtexCAD', ohlc.frequency='days')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://api.bitcoincharts.com/v1/trades.csv?symbol=SYMBOL[&end=UNIXTIME]}}
}

