% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_most_recent_trade}
\alias{get_most_recent_trade}
\title{Utility function which returns the most recent OHLC candle}
\usage{
get_most_recent_trade(symbol, ohlc.frequency = "hours", align = TRUE,
  fill = FALSE, data.directory = file.path(system.file("extdata",
  "market-data", mustWork = TRUE, package = "bitcoinchartsr"), symbol),
  debug = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the \code{'get_symbol_listing()'} method.}

\item{ohlc.frequency}{character. Supported values are \code{seconds}, \code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}

\item{debug}{logical. Debugging flag.}
}
\description{
Utility function which returns the most recent OHLC candle from the bitcoincharts.com API
}
\examples{
\dontrun{
# Get most recent 1-day candle from cavirtex.com:
get_most_recent_trade('virtexCAD', ohlc.frequency='days')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://bitcoincharts.com/markets}}
}

