% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_ohlcv_from_api}
\alias{get_ohlcv_from_api}
\title{Get OHLC for last two days of trades via API}
\usage{
get_ohlcv_from_api(symbol, start.date = as.character(now() - days(1)),
  end.date = as.character(now()), data.directory = system.file("extdata",
  mustWork = TRUE, package = "bitcoinchartsr"), ohlc.frequency = "hours",
  align = TRUE, fill = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the
\code{'get_symbol_listing()'} method.}

\item{start.date}{}

\item{end.date}{}

\item{data.directory}{character. Destination directory for downloaded data
files. Defaults to package install extdata directory.}

\item{ohlc.frequency}{character. Supported values are \code{seconds},
\code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{align}{logical. Align time series index.}

\item{fill}{logical. Fill missing values.}
}
\description{
Utility function which returns the most recent OHLC candle from
the bitcoincharts.com API
}
\examples{
\dontrun{
# Get most recent 1-day candle from cavirtex.com:
get_most_recent_trade('virtexCAD', ohlc.frequency='days')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://api.bitcoincharts.com/v1/trades.csv?symbol=SYMBOL[&end=UNIXTIME]}}
}

