% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_ohlcv_from_file}
\alias{get_ohlcv_from_file}
\title{Query bitcoincharts.com API}
\usage{
get_ohlcv_from_file(symbol, start.date = as.character(now() - days(1)),
  end.date = as.character(now()), ohlc.frequency = "hours", align = TRUE,
  fill = FALSE, data.directory = file.path(system.file("extdata", mustWork =
  TRUE, package = "bitcoinchartsr")), overwrite = FALSE)
}
\arguments{
\item{symbol}{character. Supported exchanges can be obtained by calling the
\code{'get_symbol_listing()'} method.}

\item{start.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format
representing the start of the requested data series. Defaults to 30 days prior
to current date.}

\item{end.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format
representing the start of the requested data series. Defaults to current date + 1.}

\item{ohlc.frequency}{character. Supported values are \code{seconds},
\code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{align}{logical. Align time series index.}

\item{fill}{logical. Fill missing values.}

\item{data.directory}{character. Destination directory for downloaded data
files. Defaults to package install extdata directory.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}

\item{env}{character. Environment to auto.assign the return value to.
Defaults to .GlobalEnv}
}
\description{
Query bitcoincharts.com API to obtain market data for bitcoin exchanges
}
\examples{
\dontrun{
# Get one month of hourly market data for virtexCAD:
get_ohlcv_from_file('virtexCAD')
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}

