% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_symbol_listing}
\alias{get_symbol_listing}
\title{Get list of all currently available market symbols}
\usage{
get_symbol_listing(debug = FALSE)
}
\arguments{
\item{debug}{logical. Debugging flag.}
}
\description{
Get list of all currently available market symbols from
http://bitcoincharts.com/markets
}
\examples{
\dontrun{
# Get all market symbols:
get_symbol_listing()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}
\seealso{
\code{\link{http://bitcoincharts.com/markets}}
}

