% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_ticks_from_api}
\alias{get_ticks_from_api}
\title{get_ticks_from_api}
\usage{
get_ticks_from_api(symbol)
}
\arguments{
\item{symbol}{}
}
\value{
data.frame with latest ticks from api
}
\description{
get_ticks_from_api
}
\details{
get_ticks_from_api
}

