% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_ticks_from_file}
\alias{get_ticks_from_file}
\title{get_ticks_from_file}
\usage{
get_ticks_from_file(symbol, start.date = as.character(now() - days(1)),
  end.date = as.character(now()),
  data.directory = file.path(system.file("extdata", mustWork = TRUE, package =
  "bitcoinchartsr")))
}
\arguments{
\item{symbol}{bitcoincharts symbol to get trade data from file for}

\item{start.date}{defaults to 1 day before the current time}

\item{end.date}{deafults to current time}

\item{data.directory}{directory for data file}
}
\description{
Download all historical trade data for a given symbol from
bitcoincharts.com. Then parse the files and return tick data for the specified
time period
}

