% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_ticks_from_files}
\alias{get_ticks_from_files}
\title{get_ticks_from_files}
\usage{
get_ticks_from_files(symbols = get_symbol_listing(),
  start.date = as.character(now() - days(1)),
  end.date = as.character(now()), data.directory = system.file("extdata",
  mustWork = TRUE, package = "bitcoinchartsr"))
}
\arguments{
\item{symbols}{symbols to load ticks for}

\item{start.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format
representing the start of the requested data series. Defaults to 30 days prior
to current date.}

\item{end.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format
representing the start of the requested data series. Defaults to current date + 1.}

\item{data.directory}{character. Destination directory for downloaded data
files. Defaults to package install extdata/marketdata directory.}
}
\description{
load data for all known symbols into the global environment
}
\examples{
\dontrun{
Load all data available up to the current minuute on bitcoincharts.com into
the global environment
get_ticks_from_files()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}

