% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_total_mined_coins}
\alias{get_total_mined_coins}
\title{Get total mined coins from site header table}
\usage{
get_total_mined_coins()
}
\description{
Get total mined coins from site header table
}
\examples{
\dontrun{
# Get all market symbols:
get_total_mined_coins()
}
}
\references{
\url{http://bitcoincharts.com}
}

