% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{get_trade_data}
\alias{get_trade_data}
\title{get_trade_data}
\usage{
get_trade_data(symbol, start.date = as.character(as.Date(now() - days(30))),
  end.date = as.character(as.Date(now())),
  data.directory = file.path(system.file("extdata", "market-data", mustWork =
  TRUE, package = "bitcoinchartsr"), symbol))
}
\arguments{
\item{symbol}{}

\item{start.date}{}

\item{end.date}{}

\item{data.directory}{}
}
\description{
Download all historical trade data for a given symbol from
bitcoincharts.com. Then parse the files and return tick data for the specified
time period
}

