% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{load_all_data}
\alias{load_all_data}
\title{load data for all known symbols into the global environment}
\usage{
load_all_data(start.date = as.character(as.Date(now() - days(30))),
  end.date = as.character(as.Date(now() + days(1))),
  ohlc.frequency = "days", align = TRUE, fill = FALSE,
  data.base.dir = system.file("extdata", "market-data", mustWork = TRUE,
  package = "bitcoinchartsr"), download.data = FALSE, overwrite = FALSE,
  env = .GlobalEnv, debug = FALSE)
}
\arguments{
\item{start.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format representing the start of the requested data series. Defaults to 30 days prior to current date.}

\item{end.date}{character. Character string in YYYY-MM-DD (%Y-%m-%d) format representing the start of the requested data series. Defaults to current date + 1.}

\item{ohlc.frequency}{character. Supported values are \code{seconds}, \code{minutes}, \code{hours}, \code{days}, \code{months}, \code{years}}

\item{align}{logical. Align time series index.}

\item{fill}{logical. Fill missing values.}

\item{download.data}{logical. Whether to download a fresh copy of the data file.}

\item{overwrite}{logical. Whether to overwrite the local copy of the data file.}

\item{env}{character. Environment to auto.assign the return value to. Defaults to .GlobalEnv}

\item{debug}{logical. Debugging flag.}

\item{data.directory}{character. Destination directory for downloaded data files. Defaults to package install extdata/marketdata directory.}
}
\description{
load data for all known symbols into the global environment
}
\examples{
\dontrun{
Load all data available up to the current minuute on bitcoincharts.com into the global environment
load_all_data()
}
}
\references{
\url{http://bitcoincharts.com/about/markets-api/}
}

