% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitcoinchartsr.R
\name{to_ohlc_xts}
\alias{to_ohlc_xts}
\title{to_ohlc_xts}
\usage{
to_ohlc_xts(ttime, tprice, tvolume, fmt, align = FALSE, fill = FALSE)
}
\arguments{
\item{ttime}{}

\item{tprice}{}

\item{tvolume}{}

\item{fmt}{character. Possible formats include: "%Y%m%d %H %M %S" (seconds),
"%Y%m%d %H %M" (minutes), "%Y%m%d %H" (hours), "%Y%m%d" (daily)}

\item{align}{logical. Whether or not to align time series with align.time}

\item{fill}{logical. Whether or not to impute values for missing time periods.}
}
\description{
function which converts raw tick data to xts onject with OHLC
fill whether or nto to fill in missing values so you have a regularly spaced
time series
}

